updating to new codebase
a from-scratch re-write of the code based on a demo helpfully provided by another user on the Gamebuino forums. It also includes a sprite-free way of drawing basic 'vector' style graphics. each point is given an angle (in rads) and a distance. and lines are drawn between them.

Todo:
Drawing edges that have crossed the top of the screen on the bottom, ect, for that full looping effect. (not sure how to go about this, just yet)
Asteroids (the whole point of the game, the physics bit is done for those, fortunately)
Collision detection and lasers(proceeds to silently weep)
pixel art and intro screen so that the game looks like a proper META title.